Rename proto_deserialize_input_contract.{c,h} files
