+++ src/transaction/proto_deserialize_input_contract.h
